refactor: BigDecimal 타입 변경 & @EntityListeners 추가

+++ src/main/java/com/nayoung/exchangerateopenapitest/api/dto/account/AccountRegisterRequestDto.java
@@ -5,11 +5,13 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record AccountRegisterRequestDto (
 	@NotNull(message = "Money must not be null.")
 	@Min(value = 0, message = "Money must be at least 0.")
-	Long money,
+	BigDecimal money,
 
 	@NotNull(message = "Currency must not be null.")
 	Currency currency
